Cleanup to train/predict loops.
This makes it so the collator and dataset retrieval processes are
slightly simpler, and then uses this pattern in train and predict. It is
hoped this will help with unit testing.

I have hacked around with logging information a bit so that nothing is
logged at a time when it would be confusing.